Removed QuestionsView
